requirements.txt
asyncio
aiohttp
Pillow
requests
heroku.env